requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.0.0

# Orchestration
langgraph>=0.2.0
langchain-anthropic>=0.2.0

# LLM
anthropic>=0.30.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.1.0

# Fetcher
httpx>=0.27.0

# Normalizer
trafilatura>=1.12.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
